Refactoring of check forms. WIP.
