+++ admin/package.json
@@ -47,7 +47,7 @@
     "iconv-lite": "^0.4.19",
     "jsonwebtoken": "^7.4.3",
     "migrate-mongo": "^2.2.1",
-    "moment": "^2.19.1",
+    "moment": "^2.19.2",
     "mongoose": "4.9.6",
     "mongoose-auto-increment": "^5.0.1",
     "mongoose-uuid2": "^2.0.1",
