+++ admin/views/test-developer/test-developer-home.ejs
@@ -7,17 +7,17 @@
     </div>
 </div>
 
-<div class="grid-row" aria-describedby="lead-paragraph">
+<div class="grid-row">
     <div class="column-two-thirds">
         <h1 class="heading-xlarge" tabindex="0">Multiplication tables check for test development</h1>
         <ol class="list">
             <li>
-                <h2 class="heading-small"><a href="#">Upload and view forms</a></h2>
+                <h2 class="heading-small"><a href="/test-developer/upload-and-view-forms">Upload and view forms</a></h2>
                 <p>Add, view, remove and edit check forms</p>
             </li>
 
             <li>
-                <h2 class="heading-small"><a href="#">Assign forms to check windows</a></h2>
+                <h2 class="heading-small"><a href="/test-developer/choose-check-window">Assign forms to check windows</a></h2>
                 <p>Select uploaded forms and assign them to check windows</p>
             </li>
         </ol>
