+++ admin/views/test-developer/upload-new-form.ejs
@@ -35,10 +35,10 @@
             <% } %>
         <% } %>
 
-        <h1 class="heading-xlarge" tabindex="0">Manage check forms</h1>
+        <h1 class="heading-xlarge" tabindex="0">Upload and view forms</h1>
 
         <p class="lede" tabindex="0" id="lead-paragraph" aria-labelledby="lead-paragraph">
-            Use this page to upload CSV check forms and assign them to check windows
+            View, upload or remove check forms. Each form has a check window, which is the period when pupils will take the check.
         </p>
 
         <form id="upload-form" action="/test-developer/manage-check-forms" method="post" enctype="multipart/form-data">
